TSR/requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
torch==2.0.1
transformers==4.35.0
trimesh==4.0.5
rembg
huggingface-hub
gradio
